# eskip: add gob unmarshal benchmark