```
$ benchstat -col .name <(go test ./eskip -run=NONE -bench='BenchmarkGobUnmarshal|BenchmarkParse$' -count=10)
goos: linux
goarch: amd64
pkg: github.com/zalando/skipper/eskip
cpu: Intel(R) Core(TM) i5-8350U CPU @ 1.70GHz
    │    Parse    │            GobUnmarshal            │
    │   sec/op    │   sec/op     vs base               │
*-8   238.8m ± 1%   227.9m ± 1%  -4.57% (p=0.000 n=10)

    │    Parse     │            GobUnmarshal             │
    │   bytes/op   │   bytes/op    vs base               │
*-8   15.99Mi ± 0%   17.05Mi ± 0%  +6.62% (p=0.000 n=10)

    │    Parse     │             GobUnmarshal             │
    │     B/op     │     B/op      vs base                │
*-8   49.02Mi ± 0%   70.53Mi ± 0%  +43.88% (p=0.000 n=10)

    │    Parse    │            GobUnmarshal            │
    │  allocs/op  │  allocs/op   vs base               │
*-8   1.080M ± 0%   1.020M ± 0%  -5.52% (p=0.000 n=10)
```